Module mentat::conn

Structs

Attribute
A Mentat schema attribute has a value type and several other flags determining how assertions with the attribute are interpreted.

Conn
A mutable, safe reference to the current Mentat store.

KnownEntid
An entid that's either already in the store, or newly allocated to a tempid. TODO: we'd like to link this in some way to the lifetime of a particular PartitionMap.

StructuredMap
A pull expression expands a binding into a structure. The returned structure associates attributes named in the input or retrieved from the store with values. This association is a StructuredMap.

Enums

TypedValue
Represents a value that can be stored in a Mentat store.

ValueType
The attribute of each Mentat assertion has a :db/valueType constraining the value to a particular set. Mentat recognizes the following :db/valueType values.

Traits

Syncable

Type Definitions

Entid
Represents one entid in the entid space.