Struct mentat::entity_builder::InProgressBuilder

pub struct InProgressBuilder<'a, 'c> { /* fields omitted */ }

Methods

impl<'a, 'c> InProgressBuilder<'a, 'c>

pub fn new(in_progress: InProgress<'a, 'c>) -> InProgressBuilder<'a, 'c>

pub fn transact(self) -> (InProgress<'a, 'c>, Result<TxReport, MentatError>)

Build the terms from this builder and transact them against the current InProgress. This method always returns the InProgress -- failure doesn't imply an automatic rollback.

pub fn commit(self) -> Result<TxReport, MentatError>

Transact the contents of the builder and commit the InProgress. If any step fails, roll back. Return the TxReport.