Enum mentat::Binding

pub enum Binding {
    Scalar(TypedValue),
    Vec(Arc<Vec<Binding>>),
    Map(Arc<StructuredMap>),
}

The values bound in a query specification can be:

Vecs of structured values, for multi-valued component attributes or nested expressions.

Single structured values, for single-valued component attributes or nested expressions.

Single typed values, for simple attributes.

The Binding enum defines these three options.

Datomic also supports structured inputs; at present Mentat does not, but this type would also serve that purpose.

Note that maps are not ordered, and so Binding is neither Ord nor PartialOrd.

Variants
Scalar(TypedValue)Vec(Arc<Vec<Binding>>)Map(Arc<StructuredMap>)
Methods

impl Binding

pub fn val(self) -> Option<TypedValue>

impl Binding

pub fn is_congruent_with<T>(&self, t: T) -> bool where
T: Into<Option<ValueType>>,

Returns true if the provided type is Some and matches this value's type, or if the provided type is None.

pub fn matches_type(&self, t: ValueType) -> bool

pub fn value_type(&self) -> Option<ValueType>

impl Binding

pub fn into_known_entid(self) -> Option<KnownEntid>

pub fn into_entid(self) -> Option<i64>

pub fn into_kw(self) -> Option<Arc<Keyword>>

pub fn into_boolean(self) -> Option<bool>

pub fn into_long(self) -> Option<i64>

pub fn into_double(self) -> Option<f64>

pub fn into_instant(self) -> Option<DateTime<Utc>>

pub fn into_timestamp(self) -> Option<i64>

pub fn into_string(self) -> Option<Arc<String>>

pub fn into_uuid(self) -> Option<Uuid>

pub fn into_uuid_string(self) -> Option<String>

pub fn into_c_string(self) -> Option<*mut i8>

pub fn into_kw_c_string(self) -> Option<*mut i8>

pub fn into_uuid_c_string(self) -> Option<*mut i8>