Enum mentat::TypedValue

pub enum TypedValue {
    Ref(i64),
    Boolean(bool),
    Long(i64),
    Double(OrderedFloat<f64>),
    Instant(DateTime<Utc>),
    String(Arc<String>),
    Keyword(Arc<Keyword>),
    Uuid(Uuid),
}

Represents a value that can be stored in a Mentat store.

Variants
Ref(i64)Boolean(bool)Long(i64)Double(OrderedFloat<f64>)Instant(DateTime<Utc>)String(Arc<String>)Keyword(Arc<Keyword>)Uuid(Uuid)
Methods

impl TypedValue

pub fn is_congruent_with<T>(&self, t: T) -> bool where
T: Into<Option<ValueType>>,

Returns true if the provided type is Some and matches this value's type, or if the provided type is None.

pub fn matches_type(&self, t: ValueType) -> bool

pub fn value_type(&self) -> ValueType

pub fn typed_ns_keyword(ns: &str, name: &str) -> TypedValue

Construct a new TypedValue::Keyword instance by cloning the provided values and wrapping them in a new ValueRc. This is expensive, so this might be best limited to tests.

pub fn typed_string(s: &str) -> TypedValue

Construct a new TypedValue::String instance by cloning the provided value and wrapping it in a new ValueRc. This is expensive, so this might be best limited to tests.

pub fn current_instant() -> TypedValue

pub fn instant(micros: i64) -> TypedValue

Construct a new TypedValue::Instant instance from the provided microsecond timestamp.

impl TypedValue

pub fn into_known_entid(self) -> Option<KnownEntid>

pub fn into_entid(self) -> Option<i64>

pub fn into_kw(self) -> Option<Arc<Keyword>>

pub fn into_boolean(self) -> Option<bool>

pub fn into_long(self) -> Option<i64>

pub fn into_double(self) -> Option<f64>

pub fn into_instant(self) -> Option<DateTime<Utc>>

pub fn into_timestamp(self) -> Option<i64>

pub fn into_string(self) -> Option<Arc<String>>

pub fn into_c_string(self) -> Option<*mut i8>

pub fn into_kw_c_string(self) -> Option<*mut i8>

pub fn into_uuid_c_string(self) -> Option<*mut i8>

pub fn into_uuid(self) -> Option<Uuid>

pub fn into_uuid_string(self) -> Option<String>