Struct mentat::query::RelResult

pub struct RelResult<T> {
    pub width: usize,
    pub values: Vec<T>,
}

The result you get from a 'rel' query, like:

[:find ?person ?name
 :where [?person :person/name ?name]]

There are three ways to get data out of a RelResult:

By iterating over rows as slices. Use result.rows(). This is efficient and is recommended in two cases:

If you don't need to take ownership of the resulting values (e.g., you're comparing or making a modified clone).

When the data you're retrieving is cheap to clone. All scalar values are relatively cheap: they're either small values or Rc.

By direct reference to a row by index, using result.row(i). This also returns a reference.

By consuming the results using into_iter. This allocates short-lived vectors, but gives you ownership of the enclosed TypedValues.

Fields
width: usize values: Vec<T>
Methods

impl<T> RelResult<T>

pub fn empty(width: usize) -> RelResult<T>

pub fn is_empty(&self) -> bool

pub fn row_count(&self) -> usize

Important traits for Chunks<'a, T>
impl<'a, T> Iterator for Chunks<'a, T> type Item = &'a [T];

pub fn rows(&self) -> Chunks<T>

pub fn row(&self, index: usize) -> Option<&[T]>