Struct mentat::store::Store

pub struct Store { /* fields omitted */ }

A convenience wrapper around a single SQLite connection and a Conn. This is suitable for applications that don't require complex connection management.

Methods

impl Store

pub fn open(path: &str) -> Result<Store>

Open a store at the supplied path, ensuring that it includes the bootstrap schema.

pub fn open_empty(path: &str) -> Result<Store>

Returns a totally blank store with no bootstrap schema. Use open instead.

pub fn transact(&mut self, transaction: &str) -> Result<TxReport>

impl Store

pub fn sqlite_mut(&mut self) -> &mut Connection

Intended for use from tests.

impl Store

pub fn dismantle(self) -> (Connection, Conn)

pub fn conn(&self) -> &Conn

pub fn begin_read<'m>(&'m mut self) -> Result<InProgressRead<'m, 'm>>

pub fn begin_transaction<'m>(&'m mut self) -> Result<InProgress<'m, 'm>>

pub fn cache(&mut self, attr: &Keyword, direction: CacheDirection) -> Result<()>

pub fn register_observer(&mut self, key: String, observer: Arc<TxObserver>)

pub fn unregister_observer(&mut self, key: &String)