Struct mentat::Schema

pub struct Schema {
    pub entid_map: BTreeMap<i64, Keyword>,
    pub ident_map: BTreeMap<Keyword, i64>,
    pub attribute_map: BTreeMap<i64, Attribute>,
    pub component_attributes: Vec<i64>,
}

Represents a Mentat schema.

Maintains the mapping between string idents and positive integer entids; and exposes the schema flags associated to a given entid (equivalently, ident).

TODO: consider a single bi-directional map instead of separate ident->entid and entid->ident maps.

Fields
entid_map: BTreeMap<i64, Keyword>

Map entid->ident.

Invariant: is the inverse map of ident_map.

ident_map: BTreeMap<Keyword, i64>

Map ident->entid.

Invariant: is the inverse map of entid_map.

attribute_map: BTreeMap<i64, Attribute>

Map entid->attribute flags.

Invariant: key-set is the same as the key-set of entid_map (equivalently, the value-set of ident_map).

component_attributes: Vec<i64>

Maintain a vec of unique attribute IDs for which the corresponding attribute in attribute_map has .component == true.

Methods

impl Schema

pub fn new(
ident_map: BTreeMap<Keyword, i64>,
entid_map: BTreeMap<i64, Keyword>,
attribute_map: BTreeMap<i64, Attribute>
) -> Schema

pub fn to_edn_value(&self) -> Value

Returns an symbolic representation of the schema suitable for applying across Mentat stores.

pub fn update_component_attributes(&mut self)