Trait mentat::HasSchema

pub trait HasSchema {
    fn entid_for_type(&self, t: ValueType) -> Option<KnownEntid>;

    fn get_ident<T>(&self, x: T) -> Option<&Keyword>
    where
        T: Into<i64>;

    fn get_entid(&self, x: &Keyword) -> Option<KnownEntid>;

    fn attribute_for_entid<T>(&self, x: T) -> Option<&Attribute>
    where
        T: Into<i64>;

    fn attribute_for_ident(
        &self, 
        ident: &Keyword
    ) -> Option<(&Attribute, KnownEntid)>;

    fn is_attribute<T>(&self, x: T) -> bool
    where
        T: Into<i64>;

    fn identifies_attribute(&self, x: &Keyword) -> bool;

    fn component_attributes(&self) -> &[i64];
}

Required Methods

fn entid_for_type(&self, t: ValueType) -> Option<KnownEntid>

fn get_ident<T>(&self, x: T) -> Option<&Keyword> where
T: Into<i64>,

fn get_entid(&self, x: &Keyword) -> Option<KnownEntid>

fn attribute_for_entid<T>(&self, x: T) -> Option<&Attribute> where
T: Into<i64>,

fn attribute_for_ident(
&self,
ident: &Keyword
) -> Option<(&Attribute, KnownEntid)>

fn is_attribute<T>(&self, x: T) -> bool where
T: Into<i64>,

Return true if the provided entid identifies an attribute in this schema.

fn identifies_attribute(&self, x: &Keyword) -> bool

Return true if the provided ident identifies an attribute in this schema.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

fn component_attributes(&self) -> &[i64]