Module mentat::conn

Structs

Conn
A mutable, safe reference to the current Mentat store.

InProgress
Represents an in-progress, not yet committed, set of changes to the store. Call commit to commit your changes, or rollback to discard them. A transaction is held open until you do so. Your changes will be implicitly dropped along with this struct.

InProgressRead
Represents an in-progress set of reads to the store. Just like InProgress, which is read-write, but only allows for reads.

Metadata
Connection metadata required to query from, or apply transactions to, a Mentat store.

Enums

CacheAction

CacheDirection

Traits

Pullable

Queryable

Syncable